Remove publishCurrent()
This is not desirable on script start, given bidirectional sync with C-Bus. If a remote has changed the Zigbee state when the script is not running then we want that reflected in the C-Bus object on start (because retained topic), which would be considered out of date and so set the C-Bus object to reality.